Copy items asynchronously, support moves instead of copies
update() and create() now support a 'transfer_ownership' argument that
causes the datastore to move the object, rather than copy it when it can.
When copies must be done, the copy is internally asynchronous when
possible to allow the datastore to continue processing requests from
clients while the copy is ongoing.