fix range query for timestamp to do numerical comparison instead of lexical (#1342)
With the current code, anything Xapian value-stored is converted to a string
before querying, so numeric types end up as decimal strings. This won't work as
expected because Xapian does lexical comparison. We need to use
xapian.sortable_serialise() on numeric values to convert them to an (internal)
format that will result in "numerical" comparison.

Changes index content format, so needs an index rebuild.

Prerequisite: #1437

Signed-off-by: Sascha Silbe <sascha@silbe.org>