don't destroy unchanged data store entries (SL#2668)
Calling datastore.get() and passing the file name back into datastore.update()
for the same entry with transfer_ownership set to False destroyed the data
file.

This was because source and destination file are identical (hard links) and we
overwrote the destination file in AsyncCopy.start() without unlinking it
first.

Fix this by unlinking the destination file if it exists.

Reported-By: Daniel Drake <dsd@laptop.org>
Signed-off-by: Sascha Silbe <silbe@activitycentral.com>
Acked-by: Simon Schampijer <simon@laptop.org>