Ensure we return valid internal / calculated properties
The copy in the metadata storage can get corrupted, e.g. due to low level
crashes or running out of battery (see OLPC#11372 [1] for a real-life
example).

This is especially problematic for the uid property, since without it the
caller (i.e. the Journal) can't even figure out which entry to delete.

We now determine and fill in the 'filesize' and 'uid' properties in find()
and get_properties(), instead of relying on the on-disk copy in
.../metadata/{filesize,uid}.

[1] https://dev.laptop.org/ticket/11372

Reported-by: Gary Martin <garycmartin@googlemail.com>
Signed-off-by: Sascha Silbe <silbe@activitycentral.com>
Acked-by: Simon Schampijer <simon@laptop.org>