Add ds_clean flag to trigger index rebuilds, SL #2095, #2317
This gives us more complete coverage of cases where ENOSPC or
other errors are hit when creating/updating datastore entries.

Without this patch, using the Journal on ENOSPC sometimes leads
to an "empty" Journal after restart. Datastore entries exist on
disk but are not in the index (Xapian DB).

With this patch, failure to complete create/update/delete forces
an index rebuild on restart.

Signed-off-by: Martin Langhoff <martin@laptop.org>
Tested-by: Samuel Greenfeld <greenfeld@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>